chore: 초기 세팅
