+++ README.md
@@ -2,32 +2,77 @@
 
 ## 기능요구사항
 
-- [] 당첨 번호, 로또 번호 비교
-    - [x] 로또 생성
-        - [x] 구입금액 입력
-            - [x] 재입력
-    - [x] 당첨 번호, 보너스 번호 입력
-        - [x] 재입력
-    - [] 통계
-        - [x] 로또 매칭 결과 계산
-        - [x] 총금액 계산
-        - [] 수익률 계산
-    - [x] 출력
-        - [x] 로또 구입갯수 출력
-        - [x] 로또 번호 출력
-        - [x] 당첨 결과 출력
-        - [] 수익률 출력
+- [x] 당첨 번호, 로또 번호 비교
+  - [x] 로또 생성
+    - [x] 구입금액 입력
+      - [x] 재입력
+  - [x] 당첨 번호, 보너스 번호 입력
+    - [x] 재입력
+  - [x] 통계
+    - [x] 로또 매칭 결과 계산
+    - [x] 총금액 계산
+    - [] 수익률 계산
+  - [x] 출력
+    - [x] 로또 구입갯수 출력
+    - [x] 로또 번호 출력
+    - [x] 당첨 결과 출력
+    - [] 수익률 출력
 
 ## 입력 예외 상황
- - [x] 구입금액
-    - [x] 1000미만 입력 예외
-    - [x] 정수가 아닌 입력 예외
+
+- [x] 구입금액
+  - [x] 1000미만 입력 예외
+  - [x] 정수가 아닌 입력 예외
 
 - [x] 당첨 로또번호
-    - [x] 1 ~ 45 이내가 아닌 입력 예외
-    - [x] 겹치는 번호 입력 예외
+  - [x] 1 ~ 45 이내가 아닌 입력 예외
+  - [x] 겹치는 번호 입력 예외
 
 - [x] 보너스 로또번호
-    - [x] 1 ~ 45 이내가 아닌 입력 예외
-    - [x] 당첨 로또번호와 겹치는 입력 예외
- 
+  - [x] 1 ~ 45 이내가 아닌 입력 예외
+  - [x] 당첨 로또번호와 겹치는 입력 예외
+
+---
+
+# step2 - 웹 기반 로또 게임
+
+# 🎱 로또 step2
+
+## html
+
+- [ ] 구입금액 입력 필드 + 제출 버튼
+- [ ] 구매 장수 표시 영역
+- [ ] 로또 번호 목록 표시 영역
+- [ ] 당첨 번호 입력 필드 6개
+- [ ] 보너스 번호 입력 필드
+- [ ] 당첨 번호 제출 버튼
+- [ ] 당첨 통계 테이블 (등수 / 당첨 기준 / 당첨 개수)
+- [ ] 수익률 표시 영역
+- [ ] 재시작 버튼
+- [ ] 각 입력 영역별 에러 메시지 표시 영역
+
+---
+
+## Input
+
+- [ ] `readPurchaseAmount` — 구입금액 제출 이벤트를 Promise로 래핑
+- [ ] `readWinningLottoNumber` — 당첨 번호 제출 이벤트를 Promise로 래핑
+- [ ] `readBonusNumber` — 보너스 번호 제출 이벤트를 Promise로 래핑
+- [ ] `readRetry` — 재시작 버튼 클릭 이벤트를 Promise로 래핑
+- [ ] `reRead` — 에러 발생 시 에러 메시지 표시 후 재입력 유도
+
+---
+
+## Output
+
+- [ ] `printPurchaseLottoCount` — 구매 장수 DOM 렌더링
+- [ ] `printLottos` — 로또 번호 목록 DOM 렌더링
+- [ ] `printResult` — 당첨 통계 DOM 렌더링
+- [ ] `printError` — 에러 메시지 DOM 렌더링
+- [ ] `reset` — 전체 DOM 초기 상태로 리셋
+
+---
+
+## css
+
+- [] css구현
