+++ src/LottoMachine.js
