+++ src/step1-index.js
@@ -2,8 +2,10 @@
  * step 1의 시작점이 되는 파일입니다.
  * 브라우저 환경에서 사용하는 css 파일 등을 불러올 경우 정상적으로 빌드할 수 없습니다.
  */
-import App from './step1/App.js';
+import App from "./App.js";
+import Input from "./Input.js";
+import { Output } from "./Output.js";
+import { read } from "./Console.js";
 
-const app = new App();
-
-app.run();
+const app = new App(Input, Output);
+app.run().then(() => read.close());
